requirements.txt
streamlit
streamlit-extras
python-dotenv
PyPDF2
SQLAlchemy
google-generativeai